moved staffing plan files to new folder structure
